desktop/requirements.txt
pywebview>=4.3.3
packaging
